Switch from node-sass to sass (#185)
https://eaflood.atlassian.net/browse/IWTF-3902

The node-sass package is long-deprecated so it's time to switch to sass. This includes updating gulp-sass and making some very minor tweaks to the config.

This also fixes the node-http-handler namespace. This has changed from @aws-sdk/node-http-handler to @smithy/node-http-handler.

Also make sure we specificially define this as a dependency since we do require it in our code.